refactor: Removes old enrollment component.
