+++ src/app/modules/monitoring/enrollments/enrollments.service.ts
@@ -2,7 +2,7 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { environment } from 'src/environments/environment';
-import { Enrollment } from './enrollment.model';
+import { Enrollment } from '../model/enrollment.model';
 
 @Injectable({
   providedIn: 'root',
