--- src/app/modules/monitoring/group-monitoring/enrollment/enrollment.component.html
@@ -1,33 +1,0 @@
-<mat-progress-bar style="margin: 10px" mode="indeterminate" *ngIf="progressBarActive"></mat-progress-bar>
-<div fxLayout="row" style="margin: 15px 15px 5px 15px" fxLayoutAlign=" center" fxLayoutGap="10px" *ngIf="unit">
-    <span><b>{{unit.code}}</b>: {{unit.name}}</span>
-    <span fxFlex></span>
-    <button mat-mini-fab *ngIf="isAnyUnenrolled" (click)="startEnrollment()" color="primary">
-        <mat-icon matTooltip="Omogući pristup prikazanim učenicima" [matTooltipPosition]="'below'">library_add</mat-icon>
-    </button>
-    <button mat-mini-fab *ngIf="isAnyEnrolled" (click)="unenrollAll()" color="warn">
-        <mat-icon matTooltip="Ukloni pristup prikazanim učenicima" [matTooltipPosition]="'below'">block</mat-icon>
-    </button>
-</div>
-<div fxLayout="row wrap">
-    <div fxFlex="50" *ngFor="let learnerEnrollment of enrollments">
-        <mat-card style="margin: 10px" fxLayoutGap="10px" fxLayout="row">
-            <b>{{learnerEnrollment.learner.index}} {{learnerEnrollment.learner.name}} {{learnerEnrollment.learner.surname}}</b>
-            <span fxFlex></span>
-            <div *ngIf="learnerEnrollment.enrollment" fxLayoutAlign=" center" fxLayoutGap="15px">
-                <small>
-                    Status: {{learnerEnrollment.enrollment.status}}<br>
-                    <span *ngIf="learnerEnrollment.enrollment.status === 'Active'">Aktivno od: {{learnerEnrollment.enrollment.start | date: 'short'}}</span>
-                </small>
-            </div>
-            <button *ngIf="learnerEnrollment.enrollment && learnerEnrollment.enrollment.status == 'Active'" mat-mini-fab (click)="unenroll(learnerEnrollment.learner.id)" color="warn">
-                <mat-icon matTooltip="Ukloni pristup lekciji" [matTooltipPosition]="'above'"
-                   >stop</mat-icon>
-            </button>
-            <button *ngIf="!learnerEnrollment.enrollment || learnerEnrollment.enrollment.status =='Deactivated'" mat-mini-fab (click)="startEnrollment(learnerEnrollment.learner.id)" color="primary">
-                <mat-icon matTooltip="Omogući pristup lekciji" [matTooltipPosition]="'above'"
-                   >play_arrow</mat-icon>
-            </button>
-        </mat-card>
-    </div>
-</div>
