+++ src/app/modules/monitoring/monitoring.module.ts
@@ -8,7 +8,6 @@ import { NgxChartsModule } from '@swimlane/ngx-charts';
 import { MaterialModule } from 'src/app/infrastructure/material.module';
 import {AssessmentsTableComponent} from './group-monitoring/learner-progress/assessments-table/assessments-table.component';
 import {LearnerProgressComponent} from './group-monitoring/learner-progress/learner-progress.component';
-import { EnrollmentComponent } from './group-monitoring/enrollment/enrollment.component';
 import { GradingComponent } from './grading/grading.component';
 import { ProgressComponent } from './progress/progress.component';
 import { EnrollmentsComponent } from './enrollments/enrollments.component';
@@ -18,7 +17,6 @@ import { EnrollmentsComponent } from './enrollments/enrollments.component';
     GroupMonitoringComponent,
     AssessmentsTableComponent,
     LearnerProgressComponent,
-    EnrollmentComponent,
     GradingComponent,
     ProgressComponent,
     EnrollmentsComponent,
